rfc3164: Support variant which omit the hostname
The busybox syslogd does this when logging over the network.

This cannot be autodetected, but allow a caller which is aware it is dealing
with this format to prepopulate a hostname and skip that token when parsing.

Adjust the toplevel Parse to only skip a character after parseHeader if it is a
space, since parseHostname now does not always skip to the next space.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>